+++ src/app/(auth)/signup/_hooks/useSignupForm.ts
@@ -1,0 +1,47 @@
+// hooks/useSignupForm.ts
+"use client";
+
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod";
+
+// 스키마 정의
+export const formSchema = z
+  .object({
+    name: z
+      .string()
+      .min(1, "이름을 입력하세요.")
+      .max(50, "이름은 50자 이하여야 합니다.")
+      .regex(/^[가-힣a-zA-Z\s]+$/, "이름은 한글, 영문만 입력 가능합니다."),
+    email: z
+      .email("유효한 이메일을 입력해주세요.")
+      .min(1, "이메일을 입력해주세요.")
+      .max(50, "이메일은 50자 이하여야 합니다."),
+    password: z
+      .string()
+      .min(8, "비밀번호는 8자 이상이어야 합니다.")
+      .max(50, "비밀번호는 50자 이하여야 합니다."),
+    confirmPassword: z
+      .string()
+      .min(8, "비밀번호는 8자 이상이어야 합니다.")
+      .max(50, "비밀번호는 50자 이하여야 합니다."),
+  })
+  .refine((data) => data.password === data.confirmPassword, {
+    message: "비밀번호가 일치하지 않습니다.",
+    path: ["confirmPassword"],
+  });
+
+export type SignupFormSchema = z.infer<typeof formSchema>;
+
+export function useSignupForm() {
+  return useForm<SignupFormSchema>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      name: "",
+      email: "",
+      password: "",
+      confirmPassword: "",
+    },
+    mode: "onSubmit",
+  });
+}
